% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot to a file}
\usage{
save_plot(
  plot,
  filename,
  device = NULL,
  path = NULL,
  width = NA,
  height = NA,
  units = c("in", "cm", "mm", "px"),
  dpi = 300,
  ...
)
}
\arguments{
\item{plot}{Plot to save. A ggplot object, a grid object, or any plotting
call (e.g., base R) wrapped in a function (e.g., \code{function() plot(1:10)}).}

\item{filename}{File name to create on disk.}

\item{device}{Device to use. Can either be a device function (e.g. \link{png}), or
one of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp",
"svg" or "wmf" (windows only). If \code{NULL} (default), the device is guessed
based on the \code{filename} extension.}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename} are
combined to create the fully qualified file name. Defaults to the working
directory.}

\item{width, height}{Plot size in units expressed by the \code{units} argument. If
not supplied, uses the size of the current graphics device.}

\item{units}{One of the following units in which the \code{width} and \code{height}
arguments are expressed: \code{"in"}, \code{"cm"}, \code{"mm"} or \code{"px"}.}

\item{dpi}{Plot resolution.}

\item{...}{Other arguments passed on to the graphics device function, as
specified by \code{device}.}
}
\value{
A named list (returned invisibly) with elements:
\itemize{
\item \code{file}: Full file path.
\item \code{width}: Plot width.
\item \code{height}: Plot height.
\item \code{units}: Unit of \code{width} and \code{height}.
\item \code{dpi}: Plot resolution.
}
}
\description{
Adapted from \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}. A general-purpose plot-saving function that
works with any R plotting system. Accepts ggplot objects, grid objects, or
any plotting call wrapped in a function. Unlike \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}, which
only supports ggplot and grid objects, \code{save_plot()} can capture any plotting
call (e.g., base R) by wrapping it in a function (e.g., \code{function() plot(1:10)}).
}
\seealso{
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}
}
