% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx-add-img.R
\name{docx_add_img}
\alias{docx_add_img}
\title{Add an external image to a Word document}
\usage{
docx_add_img(
  x,
  src,
  width,
  height,
  units = c("in", "cm", "mm", "px"),
  dpi = 300,
  caption = NULL,
  caption_pos = c("below", "above"),
  autonum = NULL,
  fp_text = NULL,
  fp_par = NULL
)
}
\arguments{
\item{x}{An \code{officer::read_docx} object.}

\item{src}{Path to the image file.}

\item{width}{Width of the image.}

\item{height}{Height of the image.}

\item{units}{Unit for width/height. One of \code{"in"}, \code{"cm"}, \code{"mm"}, \code{"px"}.
Default is \code{"in"}.}

\item{dpi}{Plot resolution (dots per inch). Used only if \code{units = "px"}.
Default is 300.}

\item{caption}{The caption text (string).}

\item{caption_pos}{Position of the caption relative to the image. \code{"above"}
or \code{"below"}.}

\item{autonum}{An \code{officer::run_autonum} object for figure numbering.}

\item{fp_text}{An \code{officer::fp_text} object for styling caption text.}

\item{fp_par}{An \code{officer::fp_par} object for styling the paragraph.}
}
\value{
The modified \code{rdocx} object.
}
\description{
Adds an external image to an \code{officer} Word document, with options for
sizing, captioning, and auto-numbering. It automatically resizes the image
if it exceeds the document's page margins.
}
