% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaExtra-package.R
\docType{package}
\name{metaExtra-package}
\alias{metaExtra}
\alias{metaExtra-package}
\title{metaExtra: Extra Utilities for Meta-Analysis Workflows}
\description{
Provides reusable utilities for meta-analysis workflows, including forest plot saving with auto-calculated dimensions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NourEdinDarwish/metaExtra}
  \item Report bugs at \url{https://github.com/NourEdinDarwish/metaExtra/issues}
}

}
\author{
\strong{Maintainer}: Nour Edin Darwish \email{nouredindarwish@gmail.com} (\href{https://orcid.org/0009-0009-5527-4539}{ORCID})

}
\keyword{internal}
