% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaExtra-package.R
\docType{package}
\name{metaExtra-package}
\alias{metaExtra}
\alias{metaExtra-package}
\title{metaExtra: Extra Utility Functions for Meta-Analysis}
\description{
Provides extra utility functions to complement the 'meta' package. Includes functions for automatically calculating optimal dimensions and saving forest plots to disk.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NourEdinDarwish/metaExtra}
  \item \url{https://nouredindarwish.github.io/metaExtra/}
  \item Report bugs at \url{https://github.com/NourEdinDarwish/metaExtra/issues}
}

}
\author{
\strong{Maintainer}: Nour Edin Darwish \email{nouredindarwish@gmail.com} (\href{https://orcid.org/0009-0009-5527-4539}{ORCID})

}
\keyword{internal}
