% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_plot}
\alias{save_plot}
\title{Save a ggplot, grid object, or plotting function}
\usage{
save_plot(
  x,
  filename,
  device = NULL,
  path = NULL,
  width = NA,
  height = NA,
  units = c("in", "cm", "mm", "px"),
  dpi = 300,
  ...
)
}
\arguments{
\item{x}{Plot to save.
Can be a \code{ggplot} object, a \code{grob}/\code{gTree}, a list of such objects,
or a \strong{function} that produces a base R plot (e.g.,
\code{function() plot(1:10)}).}

\item{filename}{File name to create on disk.}

\item{device}{Device to use. Can either be a device function
(e.g. \link{png}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only). If
\code{NULL} (default), the device is guessed based on the \code{filename} extension.}

\item{path}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{width, height}{Plot size in units expressed by the \code{units} argument.
If not supplied, uses the size of the current graphics device.}

\item{units}{One of the following units in which the \code{width} and \code{height}
arguments are expressed: \code{"in"}, \code{"cm"}, \code{"mm"} or \code{"px"}.}

\item{dpi}{Plot resolution.}

\item{...}{Other arguments passed on to the graphics device function,
as specified by \code{device}.}
}
\value{
The file path (invisibly).
}
\description{
Adapted from \code{ggplot2::ggsave()}
Internal utility to save plots to a file. It supports \code{ggplot} objects,
\code{grid} graphics, and base R plots (wrapped in a function).
}
\seealso{
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}
}
